requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
tabulate>=0.9.0
streamlit>=1.28.0
plotly>=5.17.0
openpyxl>=3.1.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
PyYAML>=6.0.1
aiohttp>=3.9.0
diskcache>=5.6.3
